use usize instead of u32 where applicable
Early on, I tended to default to u32 for quantities such as memory
addresses) where usize to be more correct and more ergonomic. Later in
development, as I began to realize this mistake, I used an inconsistent
mix of both. Standardize on usize everywhere that it's relevant.

Exception: in the MMU code/data structures, since this part of the
kernel is very 32-bit specific.